requirements.txt
soundcard==0.4.3
pyserial>=3.5
tensorflow[and-cuda]==2.16.1
paddlepaddle-gpu==2.6.1
transformers[torch]
modelscope~=1.13.3
funasr>=1.1.2
pytorch_lightning>=2.0.0
numpy==1.26.4
librosa==0.10.1
